requirements.txt
amqp==1.4.9
ansible==2.2.0
anyjson==0.3.3
Babel==2.3.4
backports.ssl-match-hostname==3.5.0.1
billiard==3.3.0.23
bitarray==0.8.1
bnclient==0.3
celery==3.1.24
certifi==2016.9.26
cffi==1.8.3
coreapi==2.0.8
cryptography==1.5
Django==1.10.2
django-extensions==1.7.4
django-filter==0.15.2
django-guardian==1.4.6
django-rest-swagger==2.0.6
djangorestframework==3.4.7
drf-nested-routers==0.11.1
enum34==1.1.6
flower==0.9.1
futures==3.0.5
gtextfsm==0.2.1
idna==2.1
ipaddress==1.0.17
itypes==1.1.0
Jinja2==2.8
jtextfsm==0.3.1
junos-eznc==2.0.1
kombu==3.0.37
lxml==3.6.4
MarkupSafe==0.23
MySQL-python==1.2.5
napalm==1.1.0
napalm-base==0.18.0
napalm-eos==0.4.2
napalm-fortios==0.2.1
napalm-ibm==0.1.2
napalm-ios==0.3.1
napalm-iosxr==0.3.2
napalm-junos==0.4.3
napalm-nxos==0.4.0
napalm-panos==0.2.0
napalm-pluribus==0.4.1
ncclient==0.5.2
netaddr==0.7.18
netmiko==1.1.0
openapi-codec==1.1.5
pan-python==0.10.0
paramiko==2.0.2
pyang==1.7
pyangbind==0.5.9
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pycsco==0.3.5
pyeapi==0.7.0
pyfg==0.47
pyIOSXR==0.21
pyPluribus==0.3.0
pyserial==3.1.1
pytz==2016.7
PyYAML==3.12
requests==2.11.1
requests-toolbelt==0.7.0
rest-condition==1.0.2
scp==0.10.2
simplejson==3.8.2
six==1.10.0
tornado==4.2
uritemplate==3.0.0
xmltodict==0.10.2
